Restructure database mangling code
